Address prefill SWA review comments
